restructure async in test/server.coffee
We haven't tracked down exactly why, but in some cases on MacOS the
previous setup for async testing fails. It seemed to consistently work
in Docker on Mac.

We still have an unsolved puzzle in this case: 'npx mocha' hangs on
its own whereas 'npx grunt mochaTest' exits cleanly. This behavior
happens in both MacOS and in Docker for Mac.